refactoring proposal - Code organization,Improved error handling ,Concurrency ,Resource management: The class now implements AutoCloseable, allowing for better resource management using try-with-resources.Improved performance
