+++ org.monte.media.screenrecorder/src/main/java/org.monte.media.screenrecorder/org/monte/media/screenrecorder/ScreenRecorder.java
@@ -773,211 +773,6 @@ public class ScreenRecorder extends AbstractStateModel {
     }
 
     /**
-     * This runnable grabs audio samples and enqueues them into the specified
-     * BlockingQueue. This runnable must be called twice a second.
-     */
-    private static class AudioGrabber implements Runnable {
-
-        final private TargetDataLine line;
-        final private BlockingQueue<Buffer> queue;
-        final private Format audioFormat;
-        final private int audioTrack;
-        final private long startTime;
-        private volatile long stopTime = Long.MAX_VALUE;
-        private long totalSampleCount;
-        private ScheduledFuture<?> future;
-        private long sequenceNumber;
-        private float audioLevelLeft = AudioSystem.NOT_SPECIFIED;
-        private float audioLevelRight = AudioSystem.NOT_SPECIFIED;
-        private Mixer mixer;
-
-        public AudioGrabber(Mixer mixer, Format audioFormat, int audioTrack, long startTime, BlockingQueue<Buffer> queue)
-                throws LineUnavailableException {
-            this.mixer = mixer;
-            this.audioFormat = audioFormat;
-            this.audioTrack = audioTrack;
-            this.queue = queue;
-            this.startTime = startTime;
-            DataLine.Info lineInfo = new DataLine.Info(
-                    TargetDataLine.class, AudioFormatKeys.toAudioFormat(audioFormat));
-
-            if (mixer != null) {
-                line = (TargetDataLine) mixer.getLine(lineInfo);
-            } else {
-
-                line = (TargetDataLine) AudioSystem.getLine(lineInfo);
-            }
-
-            // Make sure the line is not muted
-            try {
-                BooleanControl ctrl = (BooleanControl) line.getControl(BooleanControl.Type.MUTE);
-                ctrl.setValue(false);
-            } catch (IllegalArgumentException e) {
-                // We can't unmute the line from Java
-            }
-            // Make sure the volume of the line is bigger than 0.2
-            try {
-                FloatControl ctrl = (FloatControl) line.getControl(FloatControl.Type.VOLUME);
-                ctrl.setValue(Math.max(ctrl.getValue(), 0.2f));
-            } catch (IllegalArgumentException e) {
-                // We can't change the volume from Java
-            }
-            line.open();
-            line.start();
-        }
-
-        public void setFuture(ScheduledFuture<?> future) {
-            this.future = future;
-        }
-
-        public void close() {
-            line.close();
-        }
-
-        public synchronized void setStopTime(long newValue) {
-            this.stopTime = newValue;
-        }
-
-        public synchronized long getStopTime() {
-            return this.stopTime;
-        }
-
-        @Override
-        public void run() {
-            Buffer buf = new Buffer();
-            AudioFormat lineFormat = line.getFormat();
-            buf.format = fromAudioFormat(lineFormat).append(SilenceBugKey, true);
-
-            // For even sample rates, we select a buffer size that can 
-            // hold half a second of audio. This allows audio/video interleave
-            // twice a second, as recommended for AVI and QuickTime movies.
-            // For odd sample rates, we have to select a buffer size that can hold
-            // one second of audio. 
-            int bufferSize = lineFormat.getFrameSize() * (int) lineFormat.getSampleRate();
-            if (((int) lineFormat.getSampleRate() & 1) == 0) {
-                bufferSize /= 2;
-            }
-
-            byte bdat[] = new byte[bufferSize];
-            buf.data = bdat;
-            Rational sampleRate = Rational.valueOf(lineFormat.getSampleRate());
-            Rational frameRate = Rational.valueOf(lineFormat.getFrameRate());
-            int count = line.read(bdat, 0, bdat.length);
-            if (count > 0) {
-                computeAudioLevel(bdat, count, lineFormat);
-                buf.sampleCount = count / (lineFormat.getSampleSizeInBits() / 8 * lineFormat.getChannels());
-                buf.sampleDuration = sampleRate.inverse();
-                buf.offset = 0;
-                buf.sequenceNumber = sequenceNumber++;
-                buf.length = count;
-                buf.track = audioTrack;
-                buf.timeStamp = new Rational(totalSampleCount, 1).divide(frameRate);
-
-                // Check if recording should be stopped
-                Rational stopTS = new Rational(getStopTime() - startTime, 1000);
-                if (buf.timeStamp.add(buf.sampleDuration.multiply(buf.sampleCount)).compareTo(stopTS) > 0) {
-                    // we recorded too much => truncate the buffer 
-                    buf.sampleCount = Math.max(0, (int) Math.ceil(stopTS.subtract(buf.timeStamp).divide(buf.sampleDuration).floatValue()));
-                    buf.length = buf.sampleCount * (lineFormat.getSampleSizeInBits() / 8 * lineFormat.getChannels());
-
-                    future.cancel(false);
-                }
-                if (buf.sampleCount > 0) {
-                    try {
-                        queue.put(buf);
-                    } catch (InterruptedException ex) {
-                        // nothing to do
-                    }
-                }
-                totalSampleCount += buf.sampleCount;
-            }
-        }
-
-        /**
-         * Calculates the root-mean-square average of continuous samples. For
-         * four samples, the formula looks like this:
-         * <pre>
-         * rms = sqrt( (x0^2 + x1^2 + x2^2 + x3^2) / 4)
-         * </pre> Resources:
-         * http://www.jsresources.org/faq_audio.html#calculate_power
-         *
-         * @param data
-         * @param length
-         * @param format
-         */
-        private void computeAudioLevel(byte[] data, int length, AudioFormat format) {
-            audioLevelLeft = audioLevelRight = AudioSystem.NOT_SPECIFIED;
-            if (format.getEncoding().equals(AudioFormat.Encoding.PCM_SIGNED)) {
-                switch (format.getSampleSizeInBits()) {
-                    case 8:
-                        switch (format.getChannels()) {
-                            case 1:
-                                audioLevelLeft = computeAudioLevelSigned8(data, 0, length, format.getFrameSize());
-                                break;
-                            case 2:
-                                audioLevelLeft = computeAudioLevelSigned8(data, 0, length, format.getFrameSize());
-                                audioLevelRight = computeAudioLevelSigned8(data, 1, length, format.getFrameSize());
-                                break;
-                        }
-                        break;
-                    case 16:
-                        if (format.isBigEndian()) {
-                            switch (format.getChannels()) {
-                                case 1:
-                                    audioLevelLeft = computeAudioLevelSigned16BE(data, 0, length, format.getFrameSize());
-                                    break;
-                                case 2:
-                                    audioLevelLeft = computeAudioLevelSigned16BE(data, 0, length, format.getFrameSize());
-                                    audioLevelRight = computeAudioLevelSigned16BE(data, 2, length, format.getFrameSize());
-                                    break;
-                            }
-                        } else {
-                            switch (format.getChannels()) {
-                                case 1:
-                                    break;
-                                case 2:
-                                    break;
-                            }
-                        }
-                        break;
-                }
-            }
-        }
-
-        private float computeAudioLevelSigned16BE(byte[] data, int offset, int length, int stride) {
-            double sum = 0;
-            for (int i = offset; i < length; i += stride) {
-                int value = ((data[i]) << 8) | (data[i + 1] & 0xff);
-                sum += value * value;
-            }
-            double rms = Math.sqrt(sum / ((length - offset) / stride));
-            return (float) (rms / 32768);
-        }
-
-        private float computeAudioLevelSigned8(byte[] data, int offset, int length, int stride) {
-            double sum = 0;
-            for (int i = offset; i < length; i += stride) {
-                int value = data[i];
-
-                // FIXME - The java audio system records silence as -128 instead of 0.
-                if (value != -128) {
-                    sum += value * value;
-                }
-            }
-            double rms = Math.sqrt(sum / ((length) / stride));
-            return (float) (rms / 128);
-        }
-
-        public float getAudioLevelLeft() {
-            return audioLevelLeft;
-        }
-
-        public float getAudioLevelRight() {
-            return audioLevelRight;
-        }
-    }
-
-    /**
      * Starts file writing.
      */
     private void startWriter() {
